python/pyproject.toml
[tool.poetry]
name = "changeme"
version = "0.0.1"
description = "Change me !"
authors = ["Your name <your-name@supermail.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
jinja2 = "^3.1.2"
typer = {extras = ["all"], version = "^0.9.0"}
pyproject-parser = "^0.9.1"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
ipdb = "^0.13.13"
pytest = "^7.4.2"
ruff = "^0.0.292"
mypy = "^1.5.1"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"

[tool.config]
#hello=world
